# frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "3X-UI Panel API",
    "version": "3.x",
    "description": "Programmatic interface to a 3X-UI panel. Authenticate either by logging in (cookie) or with an API token from Settings → Security → API Token (Bearer). All endpoints under /panel/api/* honour both modes — an API token is a full-admin credential, so treat it like the panel password."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current panel (basePath aware)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token from Settings → Security → API Token. Send as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nh-v2ray-panel",
        "description": "Session cookie set by POST /login. Browser-only."
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/panel/api/server/cpuHistory/{bucket}": {
      "get": {
        "tags": [
          "Server"
        ],
        "summary": "Legacy: aggregated CPU history. Use /history/cpu/:bucket instead — same data with a uniform {t, v} shape.",
        "operationId": "get_panel_api_server_cpuHistory_bucket",
        "parameters": [
          {
            "name": "bucket",
            "in": "path",
            "required": true,
            "description": "Bucket size in seconds. Allowed: 2, 30, 60, 120, 180, 300.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/server/history/{metric}/{bucket}": {
      "get": {
        "tags": [
          "Server"
        ],
        "summary": "Aggregated time-series for one metric. Returns an array of {t, v} samples covering the last ~6 hours.",
        "operationId": "get_panel_api_server_history_metric_bucket",
        "parameters": [
          {
            "name": "metric",
            "in": "path",
            "required": true,
            "description": "cpu | mem | netUp | netDown | online | load1 | load5 | load15.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bucket",
            "in": "path",
            "required": true,
            "description": "Bucket size in seconds. Allowed: 2, 30, 60, 120, 180, 300.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": [
                    {
                      "t": 1700000000,
                      "v": 12.5
                    },
                    {
                      "t": 1700000002,
                      "v": 13.1
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/server/xrayObservatoryHistory/{tag}/{bucket}": {
      "get": {
        "tags": [
          "Server"
        ],
        "summary": "Time-series of observatory probe results for one outbound tag. Same {t, v} shape as the other history endpoints.",
        "operationId": "get_panel_api_server_xrayObservatoryHistory_tag_bucket",
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "description": "Outbound tag from the observatory config.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bucket",
            "in": "path",
            "required": true,
            "description": "Bucket size in seconds. Allowed: 2, 30, 60, 120, 180, 300.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/server/stopXrayService": {
      "post": {
        "tags": [
          "Server"
        ],
        "summary": "Stop the Xray binary. All proxies go offline immediately.",
        "operationId": "post_panel_api_server_stopXrayService",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "msg": "Xray is not running"
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/server/restartXrayService": {
      "post": {
        "tags": [
          "Server"
        ],
        "summary": "Reload Xray with the current config. Typically required after structural inbound or routing changes.",
        "operationId": "post_panel_api_server_restartXrayService",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "msg": "Xray config is invalid: ..."
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/server/installXray/{version}": {
      "post": {
        "tags": [
          "Server"
        ],
        "summary": "Download and install the specified Xray version. Pass \"latest\" for the newest release.",
        "operationId": "post_panel_api_server_installXray_version",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "description": "Xray tag (e.g. v25.10.31) or \"latest\".",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/server/updateGeofile": {
      "post": {
        "tags": [
          "Server"
        ],
        "summary": "Refresh the default GeoIP / GeoSite data files. Body can include a fileName, or use the /:fileName variant.",
        "operationId": "post_panel_api_server_updateGeofile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/server/updateGeofile/{fileName}": {
      "post": {
        "tags": [
          "Server"
        ],
        "summary": "Refresh a single Geo file by filename (e.g. geoip.dat, geosite.dat).",
        "operationId": "post_panel_api_server_updateGeofile_fileName",
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "description": "Filename of the data file to refresh.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/server/logs/{count}": {
      "post": {
        "tags": [
          "Server"
        ],
        "summary": "Return the last N lines of the panel’s own log.",
        "operationId": "post_panel_api_server_logs_count",
        "parameters": [
          {
            "name": "count",
            "in": "path",
            "required": true,
            "description": "Number of trailing log lines.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "level": "info",
                "syslog": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": "2025/01/01 12:00:00 [INFO] Server started\n2025/01/01 12:00:01 [INFO] Xray is running"
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/server/xraylogs/{count}": {
      "post": {
        "tags": [
          "Server"
        ],
        "summary": "Return the last N lines of the Xray process log.",
        "operationId": "post_panel_api_server_xraylogs_count",
        "parameters": [
          {
            "name": "count",
            "in": "path",
            "required": true,
            "description": "Number of trailing log lines.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": "2025/01/01 12:00:00 rejected  vless  proxy  example.com  reason: no valid user\n2025/01/01 12:00:01 direct  freedom  ok"
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/server/getNewEchCert": {
      "post": {
        "tags": [
          "Server"
        ],
        "summary": "Generate a new ECH (Encrypted Client Hello) keypair and config list for the given SNI.",
        "operationId": "post_panel_api_server_getNewEchCert",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}